msm: kgsl: Defer memory frees done from event kthread
Event work can result in freeing memory as part of free on timestamp.
As freeing memory can take time it shouldn't be done as part of RT
thread as it can result in hogging CPU for longer duration and not
giving chance to other tasks to run. Hence defer memory frees done
from event kthread.

Change-Id: I057ce391030efe9a5795c5cffe6364c6ff732d71
Signed-off-by: Puranam V G Tejaswi <quic_pvgtejas@quicinc.com>
[dereference23: Backport to msm-4.14]
Signed-off-by: Alexander Winkowski <dereference23@outlook.com>
Signed-off-by: One4Lots <sahorkot.ontime@gmail.com>